refactor: CartCheckList 의 컴포넌트 및 책임, 함수 분리
